refactor: extract session helpers (StatusDot, PhaseBadge, relativeTime) from AppView.tsx
Move shared UI primitives (StatusDot, PhaseBadge, relativeTime, phase/status
color maps, STATUS_DOT_CONFIG) into src/components/session/helpers.tsx so
upcoming SessionCard, SessionList, SessionDetail, and Sidebar components can
import them without circular dependencies. Re-export relativeTime from
AppView.tsx to preserve existing test imports.

Aloop-Agent: build
Aloop-Iteration: 2
Aloop-Session: orchestrator-20260321-172932-issue-154-20260321-181044
