+++ aloop/cli/dashboard/src/components/session/helpers.tsx
@@ -1,0 +1,83 @@
+import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip';
+
+// ── Relative time formatting ──
+
+export function relativeTime(ts: string): string {
+  if (!ts) return '';
+  try {
+    const diff = Date.now() - new Date(ts).getTime();
+    const mins = Math.floor(diff / 60000);
+    if (mins < 1) return 'just now';
+    if (mins < 60) return `${mins}m ago`;
+    const hrs = Math.floor(mins / 60);
+    if (hrs < 24) return `${hrs}h ago`;
+    const days = Math.floor(hrs / 24);
+    return `${days}d ago`;
+  } catch { return ''; }
+}
+
+// ── Phase colors ──
+
+export const phaseColors: Record<string, string> = {
+  plan: 'bg-purple-500/20 text-purple-600 dark:text-purple-400 border-purple-500/30',
+  build: 'bg-yellow-500/20 text-yellow-600 dark:text-yellow-400 border-yellow-500/30',
+  proof: 'bg-amber-500/20 text-amber-600 dark:text-amber-400 border-amber-500/30',
+  review: 'bg-cyan-500/20 text-cyan-600 dark:text-cyan-400 border-cyan-500/30',
+};
+
+export const phaseBarColors: Record<string, string> = {
+  plan: 'bg-purple-500', build: 'bg-yellow-500', proof: 'bg-amber-500', review: 'bg-cyan-500',
+};
+
+export const phaseDotColors: Record<string, string> = {
+  plan: 'text-purple-500', build: 'text-yellow-500', proof: 'text-amber-500', review: 'text-cyan-500',
+};
+
+export const statusColors: Record<string, string> = {
+  running: 'text-green-600 dark:text-green-400',
+  exited: 'text-muted-foreground',
+  stopped: 'text-red-600 dark:text-red-400',
+  stopping: 'text-orange-600 dark:text-orange-400',
+};
+
+// ── Phase badge component ──
+
+export function PhaseBadge({ phase, small }: { phase: string; small?: boolean }) {
+  if (!phase) return null;
+  const colors = phaseColors[phase.toLowerCase()] ?? 'bg-muted text-muted-foreground border-border';
+  const size = small ? 'px-1 py-0 text-[10px]' : 'px-1.5 py-0.5 text-xs';
+  return <span className={`inline-block rounded border font-medium ${colors} ${size}`}>{phase}</span>;
+}
+
+// ── Status dot component ──
+
+export const STATUS_DOT_CONFIG: Record<string, { color: string; label: string }> = {
+  running: { color: 'bg-green-500', label: 'Running' },
+  stopped: { color: 'bg-muted-foreground/50', label: 'Stopped' },
+  exited: { color: 'bg-muted-foreground/50', label: 'Exited' },
+  unhealthy: { color: 'bg-red-500', label: 'Unhealthy' },
+  error: { color: 'bg-red-500', label: 'Error' },
+  stuck: { color: 'bg-orange-500', label: 'Stuck' },
+  unknown: { color: 'bg-muted-foreground/30', label: 'Unknown' },
+};
+
+export function StatusDot({ status, className = '' }: { status: string; className?: string }) {
+  const config = STATUS_DOT_CONFIG[status] ?? STATUS_DOT_CONFIG.unknown;
+  const label = config.label;
+
+  const dot = status === 'running' ? (
+    <span className={`relative flex h-2.5 w-2.5 shrink-0 ${className}`}>
+      <span className={`absolute inline-flex h-full w-full animate-pulse-dot rounded-full ${config.color}/70`} />
+      <span className={`relative inline-flex h-2.5 w-2.5 rounded-full ${config.color}`} />
+    </span>
+  ) : (
+    <span className={`inline-flex h-2.5 w-2.5 shrink-0 rounded-full ${config.color} ${className}`} />
+  );
+
+  return (
+    <Tooltip>
+      <TooltipTrigger asChild><span className="inline-flex">{dot}</span></TooltipTrigger>
+      <TooltipContent><p>{label}</p></TooltipContent>
+    </Tooltip>
+  );
+}
